move running and starting runstates into their own modules
